initial commit
The ionic app for pinpoint. The routes are set up and the states are
set up.

The stream functionality is present, which locates ads around you.

The ability post ads exists.

The ability to search for ads based on categories is there too.

The auth needs A LOT of work. Need to be able to login through
passport. Right now You can only register as a new user.

Need to add Facebook, google, twitter, linked in, github Oauth.